Switch to marsey.cat for Snappy /u/. (#251)
* Switch to marsey.cat for Snappy /u/.

camas is down, replacing it with search.marsey.cat.
Note that when looking for existing Snappy comments to test against,
it appears that something else with Snappy generation is broken.

Ex: /post/66263/-/1876803 puts an entire post URL in the author field.
This commit makes no attempt to fix this. TODO for later.

* Fix Snappy body /u/ extracting author from post URL.

Following up on 6c0f4a865640:
Issue was that author was being extracted from post.url, not href.
Given that the relevant code section is specifically for /u/s in the
body text of the submission, this was a problem.